+++ server/routes/product-route.js
@@ -18,45 +18,39 @@ router.get("/", (req, res) => {
     );
 });
 
-//category類別
+//category/subcategory類別
 router.get("/category", (req, res) => {
     connection.execute(
         `SELECT
-       category_name FROM category`,
-
+        category.category_id,
+        category.category_name,
+        GROUP_CONCAT(subcategory.subcategory_name) as subcategories
+        FROM category
+        JOIN subcategory ON subcategory.category_id = category.category_id
+        GROUP BY category.category_name
+        ORDER BY category.category_id; `,
         (error, result) => {
             res.json({ result });
         }
     );
 });
 
-//subcategory類別
-router.get("/category/subcategory", (req, res) => {
-    connection.execute(
-        `SELECT
-        subcategory.subcategory_name as subcategory,
-        category.category_name AS category_name
-        FROM subcategory
-        JOIN category ON category.category_id = subcategory.category_id
-        `,
 
-        (error, result) => {
-            res.json({ result });
-        }
-    );
-});
-
+//商品細節頁
 router.get("/product-detail/:product_id", (req, res) => {
     const productId = req.params.product_id; // 取得從路由參數中傳入的 product_id
     connection.execute(
         `SELECT
         products.*,
         category.category_name AS category_name,
-        subcategory.subcategory_name AS subcategory_name
+        subcategory.subcategory_name AS subcategory_name,
+        GROUP_CONCAT(product_type.type_name) AS type_names
         FROM products
         JOIN category ON category.category_id = products.category_id
         JOIN subcategory ON subcategory.subcategory_id = products.subcategory_id
-        WHERE products.product_id = ?`,
+        LEFT JOIN product_type ON product_type.product_id = products.product_id
+        WHERE products.product_id = ?
+        GROUP BY products.product_id`,
         [productId],
         (error, result) => {
             if (error) {
@@ -68,10 +62,6 @@ router.get("/product-detail/:product_id", (req, res) => {
         }
     );
 });
-
-
-
-
 
 
 module.exports = router;
